Moved ImageAnnotations to dedicated path
