+++ dash_annotate_cv/__init__.py
@@ -1,4 +1,5 @@
 from .annotate_image_labels import AnnotateImageLabelsAIO, ImageAnnotations
 from .annotation_storage import AnnotationStorage
+from .image_annotations import ImageAnnotations
 from .image_source import ImageSource
 from .label_source import LabelSource
